requirements.txt
Django>=4.2,<5.0
numpy>=1.23
scikit-learn>=1.2
joblib>=1.2
paho-mqtt>=1.6
